addscore를 float에서 int로 변경.

+++ core/src/main/java/org/fastcatsearch/ir/query/RankInfo.java
@@ -75,7 +75,7 @@ public class RankInfo {
 		return hit;
 	}
 
-	public void addScore(float add) {
+	public void addScore(int add) {
 		score += add;
 	}
